+++ freecad/plot/Panels/Labels.py
@@ -13,7 +13,7 @@ class TaskPanel:
     def __init__(self):
         self.name = "plot labels"
         self.ui = os.path.join(os.path.dirname(__file__),
-                               "../resources/ui/",
+                               "../Resources/Interface/",
                                "Labels.ui")
         self.form = Gui.PySideUic.loadUi(self.ui)
         self.skip = False
