+++ freecad/plot/Panels/Positions.py
@@ -13,7 +13,7 @@ class TaskPanel:
     def __init__(self):
         self.name = "plot positions"
         self.ui = os.path.join(os.path.dirname(__file__),
-                               "../resources/ui/",
+                               "../Resources/Interface/",
                                "Positions.ui")
         self.form = Gui.PySideUic.loadUi(self.ui)
         self.skip = False
