+++ freecad/plot/Resources/Interface/Axes.ui
