+++ freecad/plot/Resources/Interface/Labels.ui
