+++ freecad/plot/Resources/Interface/Series.ui
