--- freecad/plot/resources/Plot.qrc
@@ -1,19 +1,0 @@
-<RCC>
-    <qresource>
-        <file>../Resources/Icons/Axes.svg</file>
-        <file>../Resources/Icons/Grid.svg</file>
-        <file>../Resources/Icons/Addon.svg</file>
-        <file>../Resources/Icons/Labels.svg</file>
-        <file>../Resources/Icons/Legend.svg</file>
-        <file>../Resources/Icons/Positions.svg</file>
-        <file>../Resources/Icons/Save.svg</file>
-        <file>../Resources/Icons/Series.svg</file>
-        <file>../Resources/Locales/Base.qm</file>
-        <file>../Resources/Locales/jp.qm</file>
-        <file>ui/Axes.ui</file>
-        <file>ui/Labels.ui</file>
-        <file>ui/Positions.ui</file>
-        <file>ui/Save.ui</file>
-        <file>ui/Series.ui</file>
-    </qresource>
-</RCC>
